Finished final game function, project is "done"
This finished out the entire project. There are some things that
I am overall unhappy with, such as my ability and understanding to
properly position the elements using flexbox. If I ever come back to
this project in the future, it should only be after learning to use
flexbox significantly better to rearrange and stylize my page better.